+++ src/DGtal/helpers/Shortcuts.h
@@ -44,6 +44,7 @@
 #include <iostream>
 #include <fstream>
 #include <sstream>
+#include <tuple>
 #include <iterator>
 #include <string>
 #include "DGtal/base/Common.h"